+++ tests/testit/test-packages.R
@@ -80,7 +80,6 @@ if (identical(tolower(Sys.getenv("CI")), "true") ||  # travis
                           install = "force")
   testit::assert(identical(packageVersion("TimeWarp", lib.loc = dir2),
                            packageVersion("TimeWarp", lib.loc = dir6)))
-  detach("package:TimeWarp", unload = TRUE)
   remove.packages("TimeWarp", lib = dir2)
   remove.packages("TimeWarp", lib = dir6)
 
@@ -91,7 +90,8 @@ if (identical(tolower(Sys.getenv("CI")), "true") ||  # travis
   pkgSnapFileRes <- data.table::fread(formals("pkgSnapshot")$packageVersionFile)
   testit::assert(is.data.frame(out))
   testit::assert(file.exists(formals("pkgSnapshot")$packageVersionFile))
-  testit::assert(isTRUE(all.equal(data.table::as.data.table(out), pkgSnapFileRes)))
+  out1 <- data.table::as.data.table(out)
+  testit::assert(isTRUE(all.equal(out1, pkgSnapFileRes)))
 
   # Skip on CRAN
   dir3 <- tempdir2("test3")
@@ -148,10 +148,10 @@ testit::assert(length(out) == 0)
 # Trigger the save available.packages and archiveAvailable
 warn <- tryCatch(out <- Require("Require (>=0.0.1)", dependencies = FALSE,
                                 install = "force"),
-                 warning = function(x) x)
+                 error = function(x) x)
 warn <- tryCatch(out <- Require("Require (>=0.0.1)", dependencies = FALSE,
                                 install = "force"),
-                 warning = function(x) x)
+                 error = function(x) x)
 if (interactive()) {
   warn <- tryCatch(out <- Require("A3 (<=0.0.1)", dependencies = FALSE,
                                   install = "force"),